feat(cards): isCardsEnabled config

+++ src/server/services/card/card.service.ts
@@ -6,7 +6,7 @@ import { Request } from '@typings/http';
 import { CardDB } from './card.db';
 import { sequelize } from '@server/utils/pool';
 import { AccountService } from '../account/account.service';
-import { CardErrors, GenericErrors, UserErrors } from '@server/../../typings/Errors';
+import { BalanceErrors, CardErrors, GenericErrors, UserErrors } from '@server/../../typings/Errors';
 import i18next from '@utils/i18n';
 import {
   BlockCardInput,
@@ -43,21 +43,30 @@ export class CardService {
     this.accountService = accountService;
   }
 
+  validateCardsConfig() {
+    if (!isFrameworkIntegrationEnabled) {
+      logger.error('Could not give card to player.');
+      throw new Error('Phsyical cards are not available without FrameworkIntegration enabled.');
+    }
+
+    if (!config.frameworkIntegration?.isCardsEnabled) {
+      logger.error('Cards are not enabled in the config.');
+      throw new Error('Cards are not enabled in the config.');
+    }
+  }
+
   async getCards(req: Request<{ accountId: number }>) {
     const cards = await this.cardDB.getByAccountId(req.data.accountId);
     return cards.map((card) => card.toJSON());
   }
 
   async getInventoryCards(req: Request): Promise<InventoryCard[]> {
+    this.validateCardsConfig();
+
     const user = this.userService.getUser(req.source);
 
     if (!user) {
       throw new Error(UserErrors.NotFound);
-    }
-
-    if (!isFrameworkIntegrationEnabled) {
-      logger.error('Phsyical cards are not available without FrameworkIntegration enabled.');
-      return [];
     }
 
     const exports = getFrameworkExports();
@@ -65,6 +74,8 @@ export class CardService {
   }
 
   async getAccountByCard(req: Request<GetATMAccountInput>): Promise<GetATMAccountResponse> {
+    this.validateCardsConfig();
+
     logger.silly('Getting account by card.');
 
     const { cardId, pin } = req.data;
@@ -97,17 +108,15 @@ export class CardService {
   }
 
   async giveCard(src: number, card: Card) {
-    if (!isFrameworkIntegrationEnabled) {
-      logger.error('Could not give card to player.');
-      logger.error('Phsyical cards are not available without FrameworkIntegration enabled.');
-      return;
-    }
+    this.validateCardsConfig();
 
     const exports = getFrameworkExports();
     return exports.giveCard(src, card);
   }
 
   async blockCard(req: Request<BlockCardInput>) {
+    this.validateCardsConfig();
+
     const { cardId, pin } = req.data;
     const card = await this.cardDB.getById(cardId);
 
@@ -128,6 +137,7 @@ export class CardService {
   }
 
   async updateCardPin(req: Request<UpdateCardPinInput>): Promise<boolean> {
+    this.validateCardsConfig();
     logger.silly('Ordering new card for source:' + req.source);
 
     const { cardId, newPin, oldPin } = req.data;
@@ -154,6 +164,7 @@ export class CardService {
   }
 
   async orderPersonalCard(req: Request<CreateCardInput>): Promise<Card | null> {
+    this.validateCardsConfig();
     logger.debug('Ordering new card for source:' + req.source);
     const { accountId, paymentAccountId, pin } = req.data;
 
@@ -191,6 +202,10 @@ export class CardService {
         },
         t,
       );
+
+      if (paymentAccount.getDataValue('balance') < newCardCost) {
+        throw new Error(BalanceErrors.InsufficentFunds);
+      }
 
       this.accountService.removeMoneyByAccountNumber({
         ...req,
